+++ .gitignore
@@ -1,3 +1,17 @@
 .idea
 _junk
 venv
+.venv
+.env
+*.pyc
+__pycache__/
+*.pyo
+*.pyd
+*.egg-info/
+build/
+dist
+.pytest_cache/
+.mypy_cache/
+.coverage
+htmlcov/
+.pypirc
