+++ environment.yml
@@ -1,0 +1,10 @@
+name: nip
+channels:
+  - defaults
+dependencies:
+  - python=3.12
+  - pip
+  - pip:
+      - -e .
+      - pytest
+      - numpy
