+++ nip/main.py
@@ -1,14 +1,14 @@
 from pathlib import Path
-from typing import Union, Any, Iterable, Callable, Optional
+from typing import Any, Callable, Iterable, Optional, Union
 
 from nip.parser.parse_func import parse, parse_string
-from nip.update import update_flatten, update
+from nip.update import update, update_flatten
+
 from . import elements
 from .constructor import Constructor
 from .convertor import Convertor
 from .dumper import Dumper
 from .non_seq_constructor import NonSequentialConstructor
-from .dict import DictObject
 
 __all__ = [
     "parse",
@@ -54,16 +54,25 @@ def construct(
     """
     if nonsequential or base_config is not None:
         base_config = base_config or config._get_root()
-        constructor = NonSequentialConstructor(base_config, strict_typing=strict_typing, as_dictobj=as_dictobj)
+        constructor = NonSequentialConstructor(
+            base_config, strict_typing=strict_typing, as_dictobj=as_dictobj
+        )
     else:
         constructor = Constructor(strict_typing=strict_typing, as_dictobj=as_dictobj)
     result = constructor.construct(config)
     return result
 
 
-def _iter_load(configs, strict_typing, nonsequential, as_dictobj):  # Otherwise load() will always be an iterator
+def _iter_load(
+    configs, strict_typing, nonsequential, as_dictobj
+):  # Otherwise load() will always be an iterator
     for config in configs:
-        yield construct(config, strict_typing=strict_typing, nonsequential=nonsequential, as_dictobj=as_dictobj)
+        yield construct(
+            config,
+            strict_typing=strict_typing,
+            nonsequential=nonsequential,
+            as_dictobj=as_dictobj,
+        )
 
 
 def load(
@@ -96,7 +105,9 @@ def load(
     if isinstance(config, Iterable):
         return _iter_load(config, strict, nonsequential, as_dictobj)
 
-    return construct(config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj)
+    return construct(
+        config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj
+    )
 
 
 def load_string(
@@ -129,7 +140,9 @@ def load_string(
     if isinstance(config, Iterable):
         return _iter_load(config, strict, nonsequential, as_dictobj)
 
-    return construct(config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj)
+    return construct(
+        config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj
+    )
 
 
 def dump(path: Union[str, Path], obj: Union[elements.Node, object]):
@@ -143,7 +156,9 @@ def dump(path: Union[str, Path], obj: Union[elements.Node, object]):
         Read or generated config if Element. In case of any other object `convert` will be called.
     """
     if not isinstance(obj, elements.Node):
-        obj = convert(obj)  # mb: wrap with Document to ensure getting `---` at the beginning of the file.
+        obj = convert(
+            obj
+        )  # mb: wrap with Document to ensure getting `---` at the beginning of the file.
     dumper = Dumper()
     dumper.dump(path, obj)
 
@@ -206,7 +221,15 @@ def _run_return(value, config, return_values, return_configs):
 
 
 def _single_run(
-    config, func, verbose, return_values, return_configs, config_parameter, strict, nonsequential, as_dictobj
+    config,
+    func,
+    verbose,
+    return_values,
+    return_configs,
+    config_parameter,
+    strict,
+    nonsequential,
+    as_dictobj,
 ):
     if verbose:
         print("=" * 20)
@@ -214,16 +237,24 @@ def _single_run(
         print(dump_string(config))
         print("----")
 
-    value = construct(config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj)
+    value = construct(
+        config, strict_typing=strict, nonsequential=nonsequential, as_dictobj=as_dictobj
+    )
     if func is not None:
-        if isinstance(value, tuple) and isinstance(value[0], list) and isinstance(value[1], dict):
+        if (
+            isinstance(value, tuple)
+            and isinstance(value[0], list)
+            and isinstance(value[1], dict)
+        ):
             args, kwargs = value
         elif isinstance(value, list):
             args, kwargs = value, {}
         elif isinstance(value, dict):
             args, kwargs = [], value
         else:
-            raise RuntimeError("Value constructed by the config cant be parsed as args and kwargs")
+            raise RuntimeError(
+                "Value constructed by the config cant be parsed as args and kwargs"
+            )
 
         if config_parameter:
             if config_parameter in kwargs:
@@ -244,11 +275,27 @@ def _single_run(
 
 
 def _iter_run(
-    configs, func, verbose, return_values, return_configs, config_parameter, strict, nonsequential, as_dictobj
+    configs,
+    func,
+    verbose,
+    return_values,
+    return_configs,
+    config_parameter,
+    strict,
+    nonsequential,
+    as_dictobj,
 ):
     for config in configs:
         run_return = _single_run(
-            config, func, verbose, return_values, return_configs, config_parameter, strict, nonsequential, as_dictobj
+            config,
+            func,
+            verbose,
+            return_values,
+            return_configs,
+            config_parameter,
+            strict,
+            nonsequential,
+            as_dictobj,
         )
         if run_return:
             yield run_return
