+++ nip/tokens.py
@@ -44,7 +44,7 @@ class Number(Token):
             try:
                 value = t(string)
                 break
-            except:
+            except (ValueError, OverflowError):
                 pass
         if value is not None:
             return len(string), Number(value)
@@ -106,7 +106,9 @@ class String(Token):
             return pos, String(stream[1 : pos - 1])
 
         pos = len(stream)
-        for op in ["#"]:  # mb: other operators stops string. (hard to raise good exception)
+        for op in [
+            "#"
+        ]:  # mb: other operators stops string. (hard to raise good exception)
             found_pos = stream.find(op)
             if found_pos >= 0:
                 pos = min(pos, found_pos)
@@ -123,14 +125,33 @@ class Name(String):
         if not (stream[pos] == "_" or stream[pos].isalpha()):  # mb: allow numeric names
             return 0, None
 
-        while pos < len(stream) and (stream[pos].isalnum() or stream[pos] == "_" or stream[pos] == "."):
+        while pos < len(stream) and (
+            stream[pos].isalnum() or stream[pos] == "_" or stream[pos] == "."
+        ):
             pos += 1
 
         return pos, Name(stream[:pos])
 
 
 class Operator(Token):
-    operators = ["---", "@", "#", "&", "!&", "!!", "!", "- ", ": ", "*", "{", "}", "[", "]", "(", ")"]
+    operators = [
+        "---",
+        "@",
+        "#",
+        "&",
+        "!&",
+        "!!",
+        "!",
+        "- ",
+        ": ",
+        "*",
+        "{",
+        "}",
+        "[",
+        "]",
+        "(",
+        ")",
+    ]
 
     @staticmethod
     def read(stream: str) -> Tuple[int, Union[None, Operator]]:
@@ -219,7 +240,9 @@ class InlinePython(Token):
 
 class PythonString(Token):
     @classmethod
-    def read(cls, stream: str, implicit_fstrings: bool = False) -> Tuple[int, Union[None, PythonString]]:
+    def read(
+        cls, stream: str, implicit_fstrings: bool = False
+    ) -> Tuple[int, Union[None, PythonString]]:
         string = stream[:].strip()
         if implicit_fstrings and string[0] in "\"'":
             if string[-1] != string[0]:
