+++ tests/base_tests/numbers/test_numbers.py
@@ -1,18 +1,28 @@
 def test_numbers():
     from nip import load
+
     res = load("base_tests/numbers/configs/numbers.nip")
-    assert res['small_number'] == 213
-    assert res['big_number'] == 308432049871029384612384762130
-    assert res['huge_number'] == 21093486304876124023441923746102384712309487123049827340921384761230948721340192384715987338581438161655146683641674865548633466485439541345343465289764254657848565684865368843566578423465642456365847434596874539788567483656875424639687446398773554672123472248137658947390568483685344663284634556745631755624716352845642368644536562741356574366285462345652217346558744566352845641345625473556874264355642345662571536755673557632474573567356783554731452624328753685241693654684783356123322511
+    assert res["small_number"] == 213
+    assert res["big_number"] == 308432049871029384612384762130
+    assert (
+        res["huge_number"]
+        == 21093486304876124023441923746102384712309487123049827340921384761230948721340192384715987338581438161655146683641674865548633466485439541345343465289764254657848565684865368843566578423465642456365847434596874539788567483656875424639687446398773554672123472248137658947390568483685344663284634556745631755624716352845642368644536562741356574366285462345652217346558744566352845641345625473556874264355642345662571536755673557632474573567356783554731452624328753685241693654684783356123322511  # noqa: E501
+    )
 
-    assert res['small_negative'] == -123
-    assert res['big_negative'] == -10230123984720394862342394
-    assert res['huge_negative'] == -128937461230847612309847123049812374019238471203948712305986235005982374509835639482756349857643058716340598743205983465423987564398756105879634056731509617508512012649087246938147565102857634958736
+    assert res["small_negative"] == -123
+    assert res["big_negative"] == -10230123984720394862342394
+    assert (
+        res["huge_negative"]
+        == -128937461230847612309847123049812374019238471203948712305986235005982374509835639482756349857643058716340598743205983465423987564398756105879634056731509617508512012649087246938147565102857634958736  # noqa: E501
+    )
 
-    assert res['small_float'] == 1.3
-    assert res['big_float'] == 43.402983470923847109487
-    assert res['huge_float'] == 2532423.40923184710239487123094821374091238471209348123740192387420394872150981670923847209348724105987204982374023984712
+    assert res["small_float"] == 1.3
+    assert res["big_float"] == 43.402983470923847109487
+    assert (
+        res["huge_float"]
+        == 2532423.40923184710239487123094821374091238471209348123740192387420394872150981670923847209348724105987204982374023984712  # noqa: E501
+    )
 
-    assert res['small_e'] == 1e2
-    assert res['big_e'] == 123e3873
-    assert res['huge_e'] == 32.323123213298e-323403
+    assert res["small_e"] == 1e2
+    assert res["big_e"] == 123e3873
+    assert res["huge_e"] == 32.323123213298e-323403
