+++ tests/features/class_construction/test_class_const.py
@@ -6,6 +6,10 @@ def test_class_const():
     result = nip.load("features/class_construction/configs/class_const_config.nip")
     expected_result = {
         "var": 123,
-        "main": {"just": "some_stuff", "simple_class": builders.SimpleClass, "def_class": builders.ClassWithDefaults},
+        "main": {
+            "just": "some_stuff",
+            "simple_class": builders.SimpleClass,
+            "def_class": builders.ClassWithDefaults,
+        },
     }
     assert result == expected_result
