+++ tests/features/nothing/test_nothing.py
@@ -6,7 +6,12 @@ def test_nothing():
 
     result = load("features/nothing/configs/nothing.nip", as_dictobj=False)
     expected_result = (
-        [2, 213, {"qwe": [3, NOTHING, NOTHING, {"fgh": ([NOTHING], {"fgr": NOTHING})}]}, NOTHING],
+        [
+            2,
+            213,
+            {"qwe": [3, NOTHING, NOTHING, {"fgh": ([NOTHING], {"fgr": NOTHING})}]},
+            NOTHING,
+        ],
         {"qwe": {"asd": {"dfg": NOTHING}}, "zxc": NOTHING, "fff": NOTHING},
     )
     assert deep_conditioned_compare(result, expected_result, [nothing_comparison])
