+++ tests/features/run/test_run.py
@@ -3,7 +3,11 @@ def test_simple_run():
     import class_example
 
     nip(class_example)
-    value = run("features/run/configs/run.nip", func=class_example.simple_class_printer, verbose=False)
+    value = run(
+        "features/run/configs/run.nip",
+        func=class_example.simple_class_printer,
+        verbose=False,
+    )
     assert value == 42
 
 
@@ -13,7 +17,10 @@ def test_inter_run():
 
     nip(builders)
     value, dumped = run(
-        "features/run/configs/run_config_param.nip", func=builders.main, config_parameter="config", verbose=False
+        "features/run/configs/run_config_param.nip",
+        func=builders.main,
+        config_parameter="config",
+        verbose=False,
     )
     assert value == "some parameter value from main with love"
     assert dumped == '---\nparam: "some parameter value"'
